% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{audit_missing}
\alias{audit_missing}
\title{This function plots the missing values in a dataset. If a value is missing in a column
for a given row then a bar is displayed.}
\usage{
audit_missing(data, na_type = NULL, order = FALSE)
}
\arguments{
\item{data}{a data.frame}

\item{na_type}{a character vector of strings that will be interpreted as NA}

\item{order}{logical, whether to order the columns and rows to display the missing
values next to each other, defautl to FALSE.}
}
\description{
This function plots the missing values in a dataset. If a value is missing in a column
for a given row then a bar is displayed.
}
\details{
If the order argument is set to TRUE then the order of the rows is modified
  and you cannot use the plot's x-axis to find a missing value in the dataset.
}
