% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_quality.R
\name{audit_numeric}
\alias{audit_numeric}
\title{This function performs a quality analysis on numeric data. It outputs a table with the
decile values.}
\usage{
audit_numeric(data, numeric_var)
}
\arguments{
\item{data}{a data.frame}

\item{numeric_var}{a character vector identifying the numeric columns to analysed in
data.}
}
\value{
a table with as many rows as numeric variables and one column per decile.
}
\description{
This function performs a quality analysis on numeric data. It outputs a table with the
decile values.
}
