% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{audit_report_excel}
\alias{audit_report_excel}
\title{Performs a quality audit of a table}
\usage{
audit_report_excel(
  data = NULL,
  quality_res = NULL,
  file = NULL,
  numeric_cutoff = -1,
  na_type = NULL,
  max_length = Inf,
  global_only = FALSE,
  na_threshold = c(40, 80),
  verbose = TRUE
)
}
\arguments{
\item{data}{a data.frame}

\item{quality_res}{an object with class qualityResult obtained with \code{data_quality()}}

\item{file}{output file name}

\item{numeric_cutoff}{the minimum number of distinct values required for a numeric
vector not to be coerced to a fator. -1 is the default, meaning no minimum required.}

\item{na_type}{charcater vector with valus that should be considered NA. Default to
NULL, no values other than regular NA are treated as NA.}

\item{max_length}{the maximum number of rows in the frequency tables}

\item{global_only}{logical, whether to return only the global summary}

\item{na_threshold}{numeric vector of length 2 defining the range of colors in the
output for the percentage of missing values. Default to c(40, 80).}

\item{verbose}{logical, should information messages be printed in the console? default to TRUE.}
}
\value{
invisible, a list with a global summary, and if available, information on numeric,
  categorical and date variables
}
\description{
This function builds an excel report based on the result of a quality check. It renders
an excel report with predefined styles using the openxlsx package.
}
\details{
If quality_res is provided, data, numeric_cutoff, na_type and max_length are ignored.
}
\examples{
data(mtcars)
report_data_quality(mtcars, file = "quality_mtcars.xlsx")

data(iris)
report_data_quality(mtcars, file = "quality_iris.xlsx")

}
