% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{audit_report_html_global}
\alias{audit_report_html_global}
\title{This function renders the data quality rmarkdown document with specific parameters.}
\usage{
audit_report_html_global(
  data,
  output_dir = ".",
  output_file = NULL,
  na_type = NULL,
  numeric_cutoff = -1,
  na_threshold = c(40, 80),
  order = FALSE
)
}
\arguments{
\item{data}{the dataset to analyse}

\item{output_dir}{the directory to write the output file to, default to the current
directory.}

\item{output_file}{name of the output file. If NULL, the default then it is
'quality_report'.}

\item{na_type}{a character vector of strings that will be interpreted as NA}

\item{na_threshold}{numeric vector defining the range of values for the percentage of
missing values to be colored green, orange and red. Default to green before 40
percent, orange between 40 and 80 and red over 80 percent. If NULL then no colors are
applied}

\item{order}{logical, whether to order the columns and rows to display the missing
values next to each other, defautl to FALSE.}
}
\description{
This function renders the data quality rmarkdown document with specific parameters.
}
\examples{
data(iris)
audit_report_html_global(iris, "iris.html")

}
