% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_cat.R
\name{audit_vector_categorical}
\alias{audit_vector_categorical}
\title{This function summarises a character vector. It returns table with descriptive
statistics and plots the frequency table.}
\usage{
audit_vector_categorical(x, max_length = 15, nchar = 20, plot = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{max_length}{the maximum number of different values to display in the freqeuncy
table and the frequency plot}

\item{nchar}{the maximum number of characters to display for each value in the plot}

\item{plot}{logical, whether to plot the graph. Defaults to TRUE.}
}
\value{
a list with two tables, one with summary statistics (lenght, number of na,
  number of unique and number of duplicate) and a frequency table for the top most
  frequent max_length values. If plot is FALSE then the graph of the frequency table is
  also returned in the list. If x has 2 unique values then a stack bar chart is
  created, if it has only one then only the tables are returned.
}
\description{
This function summarises a character vector. It returns table with descriptive
statistics and plots the frequency table.
}
\details{
Missing values are removed at the beginning and statistics are computed on the
  vector without missing data.
}
