% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_num.R
\name{audit_vector_numeric}
\alias{audit_vector_numeric}
\title{This function summarises a numeric vector. It returns table with descriptive statistics
and plots the density function, a boxplot and the cumulative density function of the
vector.}
\usage{
audit_vector_numeric(x, plot = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{plot}{logical, whether to plot the graph. Defaults to TRUE.}
}
\value{
a list with 3 tables. One with summary statistics (length, number of missing,
  min, max, mean, range, inter_quartile range, sd), another ne with the top 5 lowest
  and highest values and one with some percentiles (0, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9,
  0.95, 1). If plot is true then a plot with the empirical density, a boxplot and the
  cumulative mepirical density is also returned.
}
\description{
This function summarises a numeric vector. It returns table with descriptive statistics
and plots the density function, a boxplot and the cumulative density function of the
vector.
}
\details{
Missing values are removed at the beginning and statistics are computed on the
  vector without missing data.
}
