% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_quality.R
\name{categorical_quality}
\alias{categorical_quality}
\title{This function performs a quality analysis on categorical data. It outputs a frequency
table by variable.}
\usage{
categorical_quality(data, categorical_var, max_length = Inf)
}
\arguments{
\item{data}{a data.frame}

\item{categorical_var}{a character vector identifying the categorical columns to
analysed in data.}

\item{max_length}{the maximum number of rows in the frequency tables}
}
\value{
a list as many elements as categorical variables.
}
\description{
This function performs a quality analysis on categorical data. It outputs a frequency
table by variable.
}
