% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{data_quality}
\alias{data_quality}
\title{Performs a quality audit of a table}
\usage{
data_quality(
  data,
  numeric_cutoff = -1,
  na_type = c("", " ", "NA", "NULL"),
  max_length = Inf,
  global_only = FALSE
)
}
\arguments{
\item{data}{a data.frame.}

\item{numeric_cutoff}{the minimum number of distinct values required for a numeric
vector not to be coerced to a fator. -1 is the default, meaning no minimum required.}

\item{na_type}{charcater vector with valus that should be considered NA. Default to
NULL, no values other than regular NA are treated as NA.}

\item{max_length}{the maximum number of rows in the frequency tables}

\item{global_only}{logical, whether to return only the global summary}
}
\value{
a list with a global summary, and if available, information on numeric,
  categorical and date variables
}
\description{
This function performs a quality check on a table. The number of missing values by
variable along with the quantiles for the numeric variables and a frequency table for
each categorical variable can be found in the result.
}
\details{
The types are defined based on the types in the input table and on the value of other
arguments. 'numeric_cutoff' allows numeric variables to be classified as categorical if
they have less unique values than the value of 'numeric_cutoff'. Date, POSIXct and
POSIXlt are the only classes treated as date.
}
