% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_quality.R
\name{date_quality}
\alias{date_quality}
\title{This function performs a quality analysis on categorical data. It outputs a frequency
table by variable.}
\usage{
date_quality(data, date_var, max_length = Inf)
}
\arguments{
\item{data}{a data.frame}

\item{date_var}{a character vector identifying the date columns to analysed in data.}

\item{max_length}{the maximum number of rows in the frequency tables}
}
\value{
a list with as many elements as date variables.
}
\description{
This function performs a quality analysis on categorical data. It outputs a frequency
table by variable.
}
