% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{Helper function for 'qualityCheck'}
\usage{
freq_table(x, max_length = Inf)
}
\arguments{
\item{x}{the table containing the data,}

\item{max_length}{the maximum number of rows in the output.}
}
\value{
a frequency table with the top 'max_length' most frequent
  modalities and the associated percentage.
}
\description{
This function creates a frequency table limited to 'max_length' rows.
}
