% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{freqTable}
\alias{freqTable}
\title{Helper function for 'qualityCheck'}
\usage{
freqTable(data, name, max_unique_out = Inf)
}
\arguments{
\item{data}{the table containing the data,}

\item{name}{the name of the variable from wich the frequency table will be
computed,}

\item{max_unique_out}{the maximum number of rows in the output.}
}
\value{
a frequency table with the top 'max_unique_out' most frequent
  modalities and the associated percentage.
}
\description{
This function creates a frequency table from the variable 'name' in the table
'data'. This frequency table is ordered and limited to 'max_unique_out' rows.
}
