% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_quality.R
\name{global_quality}
\alias{global_quality}
\title{This function performs a global quality analysis on a table. It outputs the type, the
number of missing values and unique values by variable.}
\usage{
global_quality(data, numeric_cutoff = -1, na_type = c("", " ", "NA", "NULL"))
}
\arguments{
\item{data}{a data.frame.}

\item{numeric_cutoff}{the minimum number of distinct values required for a numeric
vector not to be coerced to a fator. -1 is the default, meaning no minimum required.}

\item{na_type}{charcater vector with valus that should be considered NA. Default to
NULL, no values other than regular NA are treated as NA.}
}
\value{
a list of two elements, a table with number of missing values and other
  information by variable and a named vector with the dimension of the table.
}
\description{
This function performs a global quality analysis on a table. It outputs the type, the
number of missing values and unique values by variable.
}
