% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_type.R
\name{is_categorical}
\alias{is_categorical}
\title{This function tests if a vector is of type categorial (factor, character or boolean).}
\usage{
is_categorical(x, numeric_cutoff = -1)
}
\arguments{
\item{x}{a vector}

\item{numeric_cutoff}{the minimum number of distinct values required for a numeric
vector not to be coerced to a fator. -1 is the default, meaning no minimum required.}
}
\value{
a logical
}
\description{
This function tests if a vector is of type categorial (factor, character or boolean).
}
