% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_type.R
\name{is_date}
\alias{is_date}
\title{This function tests if a vector is of type date (Date, POSIXct and POSIXlt).}
\usage{
is_date(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical
}
\description{
This function tests if a vector is of type date (Date, POSIXct and POSIXlt).
}
