% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_date.R
\name{plot_freq_date}
\alias{plot_freq_date}
\title{This function plots the frequency of records by group of dates for every scopes of the
date vector.}
\usage{
plot_freq_date(x)
}
\arguments{
\item{x}{a date vector}
}
\value{
a frequency plot by scope organized in a grid.
}
\description{
This function plots the frequency of records by group of dates for every scopes of the
date vector.
}
\details{
The scope can be hourly, daily, monhly and/or yearly. If a date vector has
  more than two different days then it has a daily scope. The same goes for months,
  years and hours.
}
