% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_date.R
\name{plot_freq_date_group}
\alias{plot_freq_date_group}
\title{This function plots the total number of records by group. Group with no records are
plotted with a red point and a
line indicating the expected number of records by hour is also added to the graph.}
\usage{
plot_freq_date_group(x, group, title = TRUE)
}
\arguments{
\item{x}{a date vector}

\item{group}{the type of grouping, one of hour, weekday, month and year.}

\item{title}{logical, whether to add a title to the graph.}
}
\value{
the plot generated
}
\description{
This function plots the total number of records by group. Group with no records are
plotted with a red point and a
line indicating the expected number of records by hour is also added to the graph.
}
