% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_date.R
\name{plot_missing_date}
\alias{plot_missing_date}
\title{This function plots the number of records by date and displays the dates without data
between the min and max of x.}
\usage{
plot_missing_date(x)
}
\arguments{
\item{x}{a date}
}
\value{
a plot of the missing dates. Depending on the scope of the date vector it can
  be a line chart or a calendar chart.
}
\description{
This function plots the number of records by date and displays the dates without data
between the min and max of x.
}
