% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qualityCheck}
\alias{qualityCheck}
\title{Performs an audit of a table.}
\usage{
qualityCheck(data, export = TRUE, file = NULL, numeric_cutoff = -1,
  max_unique_out = 100, return = FALSE, na_threshold = c(40, 80),
  id.cols = NULL, date.cols = NULL)
}
\arguments{
\item{data}{the table to analyse,}

\item{export}{logical, should a report be created,}

\item{file}{the name of the file if export is TRUE,}

\item{numeric_cutoff}{numeric value indicating the maximum number of unique
values for a numerical variable to be classified as categorical. Default to
-1 meaning that no numerical variables will be treated as categorical,}

\item{max_unique_out}{numeric value indicating the maximum number in the
output for categorical variables,}

\item{return}{logical, should a list with the results of the quality check be
returned,}

\item{na_threshold}{numeric value indicating the range of values for good, medium and bad percentage of missing values}
}
\description{
This function performs a quality check on a table and creates an excel report
if asked. In the report one can find the number of missing values by variable
alog with the quantiles for the numeric variables and a frequency table for
each categorical variable. An excel report can be rendered with the different
values computed.
}
\details{
The excel report is composed of several sheets. A 'summary' one with
information regarding missing values for every variables, a 'numeric' sheet
with the quantiles of all numeric varaibles, a 'categorical' sheet with a
table with unique values and their frequency for each categorical variable
and finally a 'date' sheet with also a frequency table for each date variable
along with the minimum and maximum.

The types are defined based on the types in the input table and on the value
of other arguments. 'numeric_cutoff' allows numeric variables to be
classified as categorical if they have less unique values than the value of
'numeric_cutoff'. Date variables has to be given by the user in the
'date.cols' argument.
}
\section{warning}{
 For now the function does not know how to deal with Date or
  POSIXlt/POSIXct format, please coerce to character before and use
  'date.cols' to specify the date format.
}

