% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_type.R
\name{which_type}
\alias{which_type}
\title{This function gives the types of a vector. It relies on is_numeric and is_date.}
\usage{
which_type(x, numeric_cutoff = -1)
}
\arguments{
\item{x}{a vector}

\item{numeric_cutoff}{the minimum number of distinct values required for a numeric
vector not to be coerced to a fator. -1 is the default, meaning no minimum required.}
}
\value{
a character string giving the type of the input vector. Can be one of
  undefined, factor, numeric, logical and character
}
\description{
This function gives the types of a vector. It relies on is_numeric and is_date.
}
