% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addCustomCell}
\alias{addCustomCell}
\title{Helper function to add custom cell}
\usage{
addCustomCell(wb, sheet, row_index, col_index, value, cell_style)
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{a sheet in the workbook}

\item{row_index}{row index of the cell}

\item{col_index}{col index of the cell}

\item{value}{value to be added to the cell}

\item{cell_style}{style object to customize the cell}
}
\description{
Helper function to add custom cell
}
