% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addTable}
\alias{addTable}
\title{This functions add a data frame to a sheet and styles it.}
\usage{
addTable(table, sheet, start.row, start.column, col.names, colnames.style,
  col.style, n.columns)
}
\arguments{
\item{table}{the table to be added,}

\item{sheet}{the sheet where tha table will be added,}

\item{start.row}{a numeric value for the starting row,}

\item{start.column}{a numeric value for the starting column,}

\item{col.names}{logical, should the colnames be written,}

\item{colnames.style}{the style associated to the colnames,}

\item{col.style}{a list of CellStyle. If the name of the list element is the
column number, it will be used to set the style of the column. Columns of
type Date and POSIXct are styled automatically even if colSyle=NULL.}
}
\description{
It uses 'addDataFrame' from the xlsx package and add birder around the table.
}
