% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{audit_report_html}
\alias{audit_report_html}
\title{This function renders the univariate exploratory analysis rmardown document with
specific parameters.}
\usage{
audit_report_html(
  data,
  numeric_cutoff = -1,
  max_length = 15,
  nchar = 20,
  output_dir = ".",
  output_file = NULL
)
}
\arguments{
\item{data}{the dataset to analyse}

\item{nchar}{maximum number of characters displayed in the plots as level values for
categorical vectors. See details.}

\item{output_dir}{the directory to write the output file to, default to the current
directory.}

\item{output_file}{name of the output file. If NULL, the default then it is 'desc_report'.}

\item{min_unique}{the minimal number of unique values for a numeric vector to be
describe as a numeric vector and not as a categorical vector. Defaults to 15.}

\item{length_out}{number of unique values displayed in the plots and table for
categorical vectors. Defaults to 15.}
}
\description{
This function renders the univariate exploratory analysis rmardown document with
specific parameters.
}
\examples{
data(iris)
audit_report_html(iris, "iris.html")

}
