% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_report.R
\name{audit_report_html}
\alias{audit_report_html}
\title{This function renders the univariate exploratory analysis rmardown document with
specific parameters.}
\usage{
audit_report_html(
  data,
  numeric_cutoff = -1,
  max_length = 15,
  nchar = 20,
  output_dir = NULL,
  output_file = NULL
)
}
\arguments{
\item{data}{the dataset to analyse}

\item{numeric_cutoff}{the minimum number of distinct values required for a numeric
vector not to be coerced to a fator. -1 is the default, meaning no minimum required.}

\item{max_length}{the maximum number of rows in the frequency tables. Default to 15.}

\item{nchar}{maximum number of characters displayed in the plots as level values for
categorical vectors. See details.}

\item{output_dir}{the directory to write the output file to, default to the current
directory.}

\item{output_file}{name of the output file. If NULL, the default then it is 'desc_report'.}
}
\description{
This function renders the univariate exploratory analysis rmardown document with
specific parameters.
}
\examples{
\donttest{
data(iris)
audit_report_html(iris, "iris.html")
}

}
